// src/Dropdown/types.ts
import { ReactNode } from 'react';
import { Placement } from '@popperjs/core';

export interface IDropdownProps {
  anchorElement: ReactNode;
  offset?: {
    horizontal: number;
    vertical: number;
  };
  placement?: Placement;
  trigger?: 'hover' | 'click';
  open: boolean;
  toggleOpen: (open: boolean) => void;
}